+++ terminus-ui/autocomplete/testing/src/test-components.ts
@@ -8,7 +8,6 @@ import {
   FormControl,
   FormsModule,
   ReactiveFormsModule,
-  Validator,
   Validators,
 } from '@angular/forms';
 import { TsAutocompleteModule } from '@terminus/ui/autocomplete';
@@ -21,7 +20,7 @@ interface State {
   disabled?: boolean;
 }
 
-const STATES: State[] = [
+export const STATES: State[] = [
   {
     name: 'Arkansas',
     population: '2.978M',
@@ -167,7 +166,7 @@ const STATES_GROUPED: GroupedStates[] = [
     >
       <ts-option
         *ngFor="let option of states"
-        [value]="option.slug"
+        [value]="option"
         [option]="option"
         [isDisabled]="option?.disabled"
       >
@@ -198,7 +197,7 @@ export class Autocomplete {
     >
       <ts-option
         *ngFor="let option of states"
-        [value]="option.slug"
+        [value]="option"
         [option]="option"
       >
         {{ option.foo }}
@@ -222,7 +221,7 @@ export class AutocompleteRequired {
     >
       <ts-option
         *ngFor="let option of states"
-        [value]="option.name"
+        [value]="option"
         [option]="option"
         [isDisabled]="option?.disabled"
       >
@@ -234,8 +233,8 @@ export class AutocompleteRequired {
   `,
 })
 export class SeededAutocomplete {
-  public myCtrl = new FormControl(['Florida']);
   public states: State[] = STATES.slice();
+  public myCtrl = new FormControl([STATES[4]]);
   public allowMultiple = true;
   public allowDuplicates = false;
   public keepOpen = false;
@@ -266,8 +265,8 @@ export class SeededAutocomplete {
   `,
 })
 export class PassingInObjectValue {
-  public myCtrl = new FormControl([{name: 'Florida'}]);
   public states: State[] = STATES.slice();
+  public myCtrl = new FormControl([STATES[4]]);
   public allowMultiple = false;
   public allowDuplicates = false;
   public keepOpen = false;
@@ -283,7 +282,7 @@ export class PassingInObjectValue {
     >
       <ts-option
         *ngFor="let option of states"
-        [value]="option.name"
+        [value]="option"
         [option]="option"
         [isDisabled]="option?.disabled"
       >
@@ -295,8 +294,29 @@ export class PassingInObjectValue {
   `,
 })
 export class SeededNgModelAutocomplete {
-  public myModel = ['Florida'];
+  public myModel = [STATES[4]];
   public states: State[] = STATES.slice();
+}
+
+@Component({
+  template: `
+    <ts-autocomplete [formControl]="meow">
+      <ts-option
+        *ngFor="let option of states"
+        [value]="option"
+        [option]="option"
+        [isDisabled]="option?.disabled"
+      >
+        <span tsOptionDisplay>
+          {{ option.name }}
+        </span>
+      </ts-option>
+    </ts-autocomplete>
+  `,
+})
+export class SeededNonArrayAutocomplete {
+  public states: State[] = STATES.slice();
+  public meow = new FormControl(this.states[0]);
 }
 
 @Component({
@@ -308,7 +328,7 @@ export class SeededNgModelAutocomplete {
     >
       <ts-option
         *ngFor="let option of states"
-        [value]="option.name"
+        [value]="option"
         [option]="option"
         [isDisabled]="option?.disabled"
       >
@@ -333,7 +353,7 @@ export class AutocompleteAllowMultipleNoReopen {
       (opened)="wasOpened($event)"
     >
       <ts-option
-        [value]="option.name"
+        [value]="option"
         [option]="option"
         [isDisabled]="option?.disabled"
         *ngFor="let option of options"
@@ -356,7 +376,7 @@ export class Disabled {
       [allowMultiple]="true"
     >
       <ts-option
-        [value]="option.name"
+        [value]="option"
         [option]="option"
         [isDisabled]="option?.disabled"
         *ngFor="let option of options"
@@ -365,7 +385,7 @@ export class Disabled {
   `,
 })
 export class SelectOptionChange {
-  public myCtrl = new FormControl(['Texas', 'Florida']);
+  public myCtrl = new FormControl([STATES[3], STATES[4]]);
   public options: State[] = STATES.slice(0, 10);
 
   public updateOptions() {
@@ -439,7 +459,7 @@ export class CustomCompareFn {
   template: `
     <ts-autocomplete [formControl]="myCtrl">
       <ts-option
-        [value]="option.name"
+        [value]="option"
         [option]="option"
         [isDisabled]="option?.disabled"
         *ngFor="let option of items"
@@ -448,7 +468,7 @@ export class CustomCompareFn {
   `,
 })
 export class DeferOptionSelectionStream {
-  public myCtrl = new FormControl('Florida');
+  public myCtrl = new FormControl([STATES[4]]);
   public items: any[] = [];
 
   public updateOptions() {
@@ -463,7 +483,7 @@ export class DeferOptionSelectionStream {
       (queryChange)="change($event)"
     >
       <ts-option
-        [value]="option.name"
+        [value]="option"
         [option]="option"
         [isDisabled]="option?.disabled"
         *ngFor="let option of options"
@@ -472,7 +492,7 @@ export class DeferOptionSelectionStream {
   `,
 })
 export class Debounce {
-  public myCtrl = new FormControl(['Florida', 'Texas']);
+  public myCtrl = new FormControl([STATES[3], STATES[4]]);
   public options = STATES.slice();
   public change = v => { };
 }
@@ -485,7 +505,7 @@ export class Debounce {
       (queryChange)="change($event)"
     >
       <ts-option
-        [value]="option.name"
+        [value]="option"
         [option]="option"
         [isDisabled]="option?.disabled"
         *ngFor="let option of options"
@@ -494,7 +514,7 @@ export class Debounce {
   `,
 })
 export class CustomDebounce {
-  public myCtrl = new FormControl(['Florida', 'Texas']);
+  public myCtrl = new FormControl([STATES[3], STATES[4]]);
   public options = STATES.slice();
   public change = v => { };
 }
@@ -508,7 +528,7 @@ export class CustomDebounce {
       (queryChange)="change($event)"
     >
       <ts-option
-        [value]="option.name"
+        [value]="option"
         [option]="option"
         [isDisabled]="option?.disabled"
         *ngFor="let option of options"
@@ -517,7 +537,7 @@ export class CustomDebounce {
   `,
 })
 export class CustomCharacterCount {
-  public myCtrl = new FormControl(['Florida', 'Texas']);
+  public myCtrl = new FormControl([STATES[3], STATES[4]]);
   public options = STATES.slice();
   public customCount: number | undefined;
   public change = v => { };
@@ -531,7 +551,7 @@ export class CustomCharacterCount {
       [isRequired]="true"
     >
       <ts-option
-        [value]="option.name"
+        [value]="option"
         [option]="option"
         [isDisabled]="option?.disabled"
         *ngFor="let option of options"
@@ -552,7 +572,7 @@ export class HideRequired {
       [hint]="myHint"
     >
       <ts-option
-        [value]="option.name"
+        [value]="option"
         [option]="option"
         [isDisabled]="option?.disabled"
         *ngFor="let option of options"
@@ -573,7 +593,7 @@ export class Hint {
       [id]="myId"
     >
       <ts-option
-        [value]="option.name"
+        [value]="option"
         [option]="option"
         [isDisabled]="option?.disabled"
         *ngFor="let option of options"
@@ -594,7 +614,7 @@ export class Id {
       [label]="myLabel"
     >
       <ts-option
-        [value]="option.name"
+        [value]="option"
         [option]="option"
         [isDisabled]="option?.disabled"
         *ngFor="let option of options"
@@ -616,7 +636,7 @@ export class Label {
       [validateOnChange]="validateOnChange"
     >
       <ts-option
-        [value]="option.name"
+        [value]="option"
         [option]="option"
         [isDisabled]="option?.disabled"
         *ngFor="let option of options"
@@ -634,7 +654,7 @@ export class ValidateOnChange {
   template: `
     <ts-autocomplete [formControl]="myCtrl">
       <ts-option
-        [value]="option.value"
+        [value]="option"
         [option]="option"
         *ngFor="let option of items"
       >{{ option.viewValue }}</ts-option>
@@ -642,7 +662,6 @@ export class ValidateOnChange {
   `,
 })
 export class NullSelection {
-  public myCtrl = new FormControl('bar');
   public items = [
     {
       value: 'foo',
@@ -657,13 +676,14 @@ export class NullSelection {
       viewValue: 'bar view',
     },
   ];
+  public myCtrl = new FormControl(this.items[2]);
 }
 
 @Component({
   template: `
     <ts-autocomplete [formControl]="myCtrl">
       <ts-option
-        [value]="state.name"
+        [value]="state"
         *ngFor="let state of items"
       >
         <ng-template let-option>
@@ -685,7 +705,7 @@ export class OptionError {
   template: `
     <ts-autocomplete [formControl]="myCtrl">
       <ts-option
-        [value]="state.name"
+        [value]="state"
         [option]="state"
         [id]="state.name"
         *ngFor="let state of items"
@@ -712,7 +732,7 @@ export class OptionId {
       >
         <ts-option
           *ngFor="let option of group.children"
-          [value]="option.name"
+          [value]="option"
           [option]="option"
         >
           {{ option.name }}
@@ -736,7 +756,7 @@ export class OptgroupIDs {
       >
         <ts-option
           *ngFor="let option of group.children"
-          [value]="option.name"
+          [value]="option"
           [option]="option"
         >
           {{ option.name }}
@@ -746,7 +766,7 @@ export class OptgroupIDs {
   `,
 })
 export class OptgroupBadIDs {
-  public myCtrl = new FormControl('Florida');
+  public myCtrl = new FormControl([STATES[4]]);
   public groups = STATES_GROUPED.slice();
 }
 
@@ -782,6 +802,7 @@ export class OptgroupBadIDs {
     OptionError,
     OptionId,
     SeededAutocomplete,
+    SeededNonArrayAutocomplete,
     PassingInObjectValue,
     SeededNgModelAutocomplete,
     SelectOptionChange,