+++ CHANGELOG.md
@@ -1,3 +1,24 @@
+## [14.8.2](https://github.com/GetTerminus/terminus-ui/compare/v14.8.1...v14.8.2) (2019-08-22)
+
+
+### Bug Fixes
+
+* **radio-group:** fix emitting the correct TsRadioChange on changes ([e9c0d63](https://github.com/GetTerminus/terminus-ui/commit/e9c0d63))
+
+## [14.8.1](https://github.com/GetTerminus/terminus-ui/compare/v14.8.0...v14.8.1) (2019-08-22)
+
+
+### Bug Fixes
+
+* **autocomplete:** add comparator/formatter. fix things ([aeb09a8](https://github.com/GetTerminus/terminus-ui/commit/aeb09a8))
+
+# [14.8.0](https://github.com/GetTerminus/terminus-ui/compare/v14.7.0...v14.8.0) (2019-08-22)
+
+
+### Features
+
+* **Autocomplete:** chip display based on option view ([6b4cca2](https://github.com/GetTerminus/terminus-ui/commit/6b4cca2))
+
 # [14.7.0](https://github.com/GetTerminus/terminus-ui/compare/v14.6.0...v14.7.0) (2019-08-15)